Moved score_compounds from utils to metabolomics to avoid circular imports
