+++ minedatabase/metabolomics.py
@@ -7,7 +7,6 @@ import os
 import re
 import time
 import xml.etree.ElementTree as ET
-from ast import literal_eval
 from typing import Callable, Generator, List, Optional, Set, Tuple
 
 import numpy as np
@@ -15,7 +14,7 @@ import pymongo
 
 import minedatabase
 from minedatabase.databases import MINE
-from minedatabase.utils import mongo_ids_to_mine_ids, score_compounds
+from minedatabase.utils import mongo_ids_to_mine_ids
 
 
 MINEDB_DIR = os.path.dirname(minedatabase.__file__)
@@ -1228,3 +1227,66 @@ def spectra_download(db: MINE, mongo_id: str = None) -> str:
     spectral_library = "\n".join(spectral_library)
 
     return spectral_library
+
+
+def score_compounds(
+    compounds: list,
+    model_id: str = None,
+    core_db: pymongo.database = None,
+    mine_db: pymongo.database = None,
+    kegg_db: pymongo.database = None,
+    parent_frac: float = 0.75,
+    reaction_frac: float = 0.25,
+    get_native: bool = False
+) -> List[dict]:
+    """This function validates compounds against a metabolic model, returning
+    only the compounds which pass.
+
+    Parameters
+    ----------
+    db : Mongo DB
+        Should contain a "models" collection with compound and reaction IDs
+        listed.
+    core_db : Mongo DB
+        Core MINE database.
+    compounds : list
+        Each element is a dict describing that compound. Should have an '_id'
+        field.
+    model_id : str
+        KEGG organism code (e.g. 'hsa').
+    parent_frac : float, optional
+        Weighting for compounds derived from compounds in the provided model.
+        0.75 by default.
+    reaction_frac : float, optional
+        Weighting for compounds derived from known compounds not in the model.
+        0.25 by default.
+
+    Returns
+    -------
+    compounds : List[dict]
+        Modified version of input compounds list, where each compound now has
+        a 'Likelihood_score' key and value between 0 and 1.
+    """
+    for cpd in compounds:
+        cpd['native_hit'] = False
+        cpd['product_of_native_hit'] = False
+
+    if get_native and core_db and mine_db and kegg_db and model_id:
+        cpd_ids = [cpd['_id'] for cpd in compounds]
+        met_dataset = MetabolomicsDataset(name='temp')
+        met_dataset.native_set = get_KEGG_comps(mine_db, core_db, kegg_db, [model_id])
+        native_product_ids = met_dataset.check_product_of_native(cpd_ids, mine_db)
+
+        for cpd in compounds:
+            if cpd['_id'] in met_dataset.native_set:
+                cpd['native_hit'] = True
+            elif cpd['_id'] in native_product_ids:
+                cpd['product_of_native_hit'] = True
+
+    for comp in compounds:
+        if comp["native_hit"]:
+            comp["Likelihood_score"] = parent_frac + reaction_frac
+        if comp["product_of_native_hit"]:
+            comp["Likelihood_score"] = parent_frac
+
+    return compounds
