+++ minedatabase/utils.py
@@ -13,7 +13,6 @@ from numbers import Number
 from os import path
 from typing import List, Tuple, Union
 
-import pymongo
 import rdkit
 from rdkit.Chem import AllChem
 
@@ -435,51 +434,6 @@ def neutralise_charges(
             rms = AllChem.ReplaceSubstructs(mol, reactant, product)
             mol = rms[0]
     return mol
-
-
-def score_compounds(
-    db: pymongo.database,
-    compounds: list,
-    model_id: str,
-    parent_frac: float = 0.75,
-    reaction_frac: float = 0.25,
-) -> List[dict]:
-    """This function validates compounds against a metabolic model, returning
-    only the compounds which pass.
-
-    Parameters
-    ----------
-    db : Mongo DB
-        Should contain a "models" collection with compound and reaction IDs
-        listed.
-    core_db : Mongo DB
-        Core MINE database.
-    compounds : list
-        Each element is a dict describing that compound. Should have an '_id'
-        field.
-    model_id : str
-        KEGG organism code (e.g. 'hsa').
-    parent_frac : float, optional
-        Weighting for compounds derived from compounds in the provided model.
-        0.75 by default.
-    reaction_frac : float, optional
-        Weighting for compounds derived from known compounds not in the model.
-        0.25 by default.
-
-    Returns
-    -------
-    compounds : List[dict]
-        Modified version of input compounds list, where each compound now has
-        a 'Likelihood_score' key and value between 0 and 1.
-    """
-
-    for comp in compounds:
-        if comp["native_hit"]:
-            comp["Likelihood_score"] = parent_frac + reaction_frac
-        if comp["product_of_native_hit"]:
-            comp["Likelihood_score"] = parent_frac
-
-    return compounds
 
 
 def get_atom_count(
